+++ FastAnimations/Framework/ModConfig.cs
@@ -21,6 +21,9 @@ namespace Pathoschild.Stardew.FastAnimations.Framework
         /// <summary>The speed multiplier for milking.</summary>
         public float MilkSpeed { get; set; } = 5;
 
+        /// <summary>The speed multiplier for mounting or dismounting the horse.</summary>
+        public float MountOrDismountSpeed { get; set; } = 2;
+
         /// <summary>The speed multiplier for shearing.</summary>
         public float ShearSpeed { get; set; } = 5;